Remove README.md as packaged data in the wheel
README.md is not needed here except for reading to provide the
`long_description` attribute in `setup.py`, which is then used only
when uploading to pypi. It seems we can safely remove it from the
manifest while retaining it in git.

Normally the `data` published here is for non-python files necessary
for the program to function (e.g., i18n translations).